requirements.txt
rgbmatrix
Pillow
requests
click